define AsyncArray
