+++ src/array/mod.rs
@@ -1,3 +1,5 @@
+mod r#async;
 mod sync;
 
+pub use r#async::PyAsyncArray;
 pub use sync::PyArray;
