+++ src/node/mod.rs
@@ -2,5 +2,5 @@
 mod node;
 mod node_path;
 
-pub(crate) use node::open_node;
+pub(crate) use node::{open_node, open_node_async};
 pub(crate) use node_path::PyNodePath;
